telegram/requirements.txt
python-telegram-bot
pyyaml
jinja2